requirements.txt
torch>=2.1.0
stable-baselines3>=2.2.0
gymnasium>=0.29.0
numpy>=1.26.0
pandas>=2.1.0
shap>=0.44.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
boto3>=1.34.0
botocore>=1.34.0
confluent-kafka>=2.3.0
prometheus-client>=0.20.0
matplotlib>=3.8.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0